sd build and tests
* Merge config and fix Dockerfile

* Add the screw driver build file.

* Patch to remove internal dependency.

* Remove unused dependencies

* Remove openssl from Dockerfile

* Use Rust container 1.51 and install clang + openssl

* add rustfmt to build

* fix PR build and run tests

* Fix tests

* run tests during commit build

* publish docs as gh-pages

* Install docker for build

* Install docker in sd

* Release build

* Fixing publish-docs job

* Publish docs

* Fixing gh-pages

+++ src/utils/config.rs
@@ -38,7 +38,7 @@ pub struct Config {
     pub ssl_cert: String,
     pub ca_cert: String,
     pub polling_interval: u64,
-    pub segment_duration: u32,
+    pub segment_duration: u64,
     pub namespace: String,
     pub start_epoch: u64,
     pub input_bucket: String,
@@ -68,7 +68,7 @@ impl Config {
                 .get_str("data_path")
                 .unwrap_or(String::from("/var/myst/data/"))
                 .to_string(),
-            
+
             cache,
             docstore_block_size: config.get_int("docstore_block_size").unwrap_or(200) as usize,
             log_file: config
@@ -79,9 +79,9 @@ impl Config {
             ssl_key: config.get_str("ssl_key").unwrap(),
             ssl_cert: config.get_str("ssl_cert").unwrap(),
             ca_cert: config.get_str("ca_cert").unwrap(),
-            
+
             polling_interval: config.get_int("polling_interval").unwrap() as u64,
-            segment_duration: config.get_int("segment_duration").unwrap() as u32,
+            segment_duration: config.get_int("segment_duration").unwrap() as u64,
             namespace: config.get_str("namespace").unwrap().to_string(),
             start_epoch: config.get_int("start_epoch").unwrap() as u64,
             input_bucket: config.get_str("input_bucket").unwrap().to_string(),
@@ -90,7 +90,6 @@ impl Config {
             aws_secret: config.get_str("aws_secret").unwrap().to_string(),
             temp_data_path: config.get_str("temp_data_path").unwrap().to_string(),
             download_frequency: config.get_int("download_frequency").unwrap() as u64,
-            
         }
     }
 }
